requirements.txt
Flask>=3.1,<3.2
python-dotenv>=1.0,<2.0
openpyxl>=3.1,<3.2
